Move `BridgeController` types into `shared` to resolve circular dependencies
